lol c++

+++ src/Menu/OptionsModsState.cpp
@@ -352,7 +352,7 @@ void OptionsModsState::moveModDown(Action *action, unsigned int row, bool max)
 			}
 			targetScrollPos += _lstMods->getNumTextLines(i);
 		}
-		if (curScrollPos + _lstMods->getVisibleRows() > targetScrollPos)
+		if (curScrollPos + (int)_lstMods->getVisibleRows() > targetScrollPos)
 		{
 			int ydiff = _lstMods->getTextHeight(row + 1);
 			SDL_WarpMouse(action->getLeftBlackBand() + action->getXMouse(),
